Split button repeating logic from `TimingAdjustButton` to own component
